xrGame/Actor_Network.cpp: change types of the g_start_level_vertex_id and the g_start_game_vertex_id

+++ src/xrGame/Actor.h
@@ -823,8 +823,8 @@ IC CActorCondition& CActor::conditions() const
 }
 
 extern Fvector g_start_position;
-extern int g_start_game_vertex_id;
-extern int g_start_level_vertex_id;
+extern u16 g_start_game_vertex_id;
+extern u32 g_start_level_vertex_id;
 extern shared_str g_start_position_smart;
 extern CActor* g_actor;
 CActor* Actor();
